requirements.txt
fastapi
uvicorn
requests
google-generativeai
pinecone
langchain
python-dotenv
tqdm
pypdf